requirements.txt
pyyaml
psycopg2-binary
torch
transformers
requests
groq
streamlit
docker